Identified error on resize
If resizing and the width goes over something like 15 pixels more than the initial size... the vkQueueSubmit2 call in the draw function returns VK_ERROR_DEVICE_LOST.
I haven't identified the cause of this error but it seems to be invalid usage when resizing above a certain point. I can resize to smaller sizes just fine, but not over a certain width... Tested with setting a smaller width initially but it failed in the same way. I can also put the window extent to the monitor's resolution and it will go fullscreen as normal.  I guess it's workable for now but a very strange error imo... will have to dig deeper!