"fix" resizing issue by making drawimage larger
The draw image used to be hardcoded 1700x900 but it would crash if resizing window to above this res ..
ix it by making the draw image larger than the window and just using part of the draw image to render...
It's not a pretty fix, but it does the job here. "_largestExtent" is set to 2560x1440, the resolutio of my main desktop monitor... but this might make the laptop work harder than needed. At some point will need to implement a more robust and dynamic resizing system.